refactor: gerenciar componente e estilos na mesma pasta
