+++ meu-app/frontend/app/components/settingsSchedules/styles.js
